+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- nothing yet
+### Changed
+
+- Base mobile styling in place
 
 ## [v0.2.0] - 2025-10-23