+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Final styling like 99% in place
+- Final styling like 99% in place (not sure if i'm in love with the button styles yet)
 - Made the header the save name field for better mobile experience
 
 ### Fixed